arrays updated

+++ TypeConversion/src/exercise/TypeConversionOne.java
@@ -4,9 +4,9 @@ public class TypeConversionOne {
 
 	public static void main(String[] args) {
 		
-		float radius = 3;
+		int radius = 3;
 				
-		int area = (int)(radius*radius*3.14); 
+		float area = (radius*radius*3.14f); 
 		
 		System.out.println(area);
